Fixed help, refactored helper cmds
