+++ src/util/index.ts
@@ -1,10 +1,11 @@
+export * from "./cooldown-helper.js";
 export * from "./db-helper.js";
 export * from "./delete-voice-channel.js";
 export * from "./embed-helpers.js";
 export * from "./enum-helper.js";
+export * from "./error-helper.js"
 export * from "./general.js";
 export * from "./idea-helpers.js";
-export * from "./message-channel.js";
 export * from "./music-helpers.js";
 export * from "./person-theme-helpers.js";
 export * from "./print-space.js";
